Refactor 6 - BaseBoss
boss package has now been moved, within includes :
- Boss
- BossProjectile
- ISD
- ISDProjectile

boss logic has been consolidated into BaseBoss class

Shield interface has 2 new additions :
- getFitWidth
- getFitHeight

LevelISD and LevelTwo now rely on LevelView instead of LevelViewLevelTwo, handling boss related logic within themselves

LevelViewLevelTwo has been deleted due to redundancy
