--- CW2024-master/CW2024-master/src/main/java/dev/ccr/dmscw2024/Boss.java
@@ -1,156 +1,0 @@
-package dev.ccr.dmscw2024;
-
-import java.util.*;
-
-public class Boss extends FighterPlane {
-
-	private static final String IMAGE_NAME = "/dev/ccr/dmscw2024/images/bossplane.png";
-	private static final double INITIAL_X_POSITION = 1000.0;
-	private static final double INITIAL_Y_POSITION = 400;
-	private static final double PROJECTILE_Y_POSITION_OFFSET = 75.0;
-	private static final double BOSS_FIRE_RATE = .04;
-	private static final double BOSS_SHIELD_PROBABILITY = 0.015;
-	private static final int IMAGE_HEIGHT = 300;
-	private static final int VERTICAL_VELOCITY = 8;
-	private static final int HEALTH = 10;
-	private static final int MOVE_FREQUENCY_PER_CYCLE = 5;
-	private static final int ZERO = 0;
-	private static final int MAX_FRAMES_WITH_SAME_MOVE = 10;
-	private static final int Y_POSITION_UPPER_BOUND = -100;
-	private static final int Y_POSITION_LOWER_BOUND = 475;
-	private static final int MAX_FRAMES_WITH_SHIELD = 250;
-	private final List<Integer> movePattern;
-	private boolean isShielded;
-	private int consecutiveMovesInSameDirection;
-	private int indexOfCurrentMove;
-	private int framesWithShieldActivated;
-
-	private final ShieldImage shieldImage;
-
-	public Boss() {
-		super(IMAGE_NAME, IMAGE_HEIGHT, INITIAL_X_POSITION, INITIAL_Y_POSITION, HEALTH);
-		System.out.println("Boss: Base Constructor COMPLETE");
-		movePattern = new ArrayList<>();
-		consecutiveMovesInSameDirection = 0;
-		indexOfCurrentMove = 0;
-		framesWithShieldActivated = 0;
-		isShielded = false;
-		initializeMovePattern();
-
-		shieldImage = new ShieldImage(INITIAL_X_POSITION, INITIAL_Y_POSITION);
-		System.out.println("Boss: ShieldImage CREATED");
-	}
-
-	@Override
-	public void updatePosition() {
-		double initialTranslateY = getTranslateY();
-		moveVertically(getNextMove());
-
-		double currentX = getLayoutY() + getTranslateY();
-		double currentY = getLayoutX() + getTranslateX();
-
-		double currentPosition = getLayoutY() + getTranslateY();
-		if (currentPosition < Y_POSITION_UPPER_BOUND || currentPosition > Y_POSITION_LOWER_BOUND) {
-			setTranslateY(initialTranslateY);
-		}
-
-		shieldImage.setLayoutY(currentX + IMAGE_HEIGHT / 4.0);
-		shieldImage.setLayoutX(currentY - IMAGE_HEIGHT / 2.0);
-		//System.out.println("SHIELD POSITION UPDATED");
-	}
-	
-	@Override
-	public void updateActor() {
-		updatePosition();
-		updateShield();
-	}
-
-	@Override
-	public ActiveActorDestructible fireProjectile() {
-		return bossFiresInCurrentFrame() ? new BossProjectile(getProjectileInitialPosition()) : null;
-	}
-	
-	@Override
-	public void takeDamage() {
-		if (!isShielded) {
-			super.takeDamage();
-		}
-		if (getHealth() <= 0){
-			destroy();
-		}
-	}
-
-	public boolean isDestroyed(){
-		return getHealth() <= 0;
-	}
-
-	private void initializeMovePattern() {
-		for (int i = 0; i < MOVE_FREQUENCY_PER_CYCLE; i++) {
-			movePattern.add(VERTICAL_VELOCITY);
-			movePattern.add(-VERTICAL_VELOCITY);
-			movePattern.add(ZERO);
-		}
-		Collections.shuffle(movePattern);
-	}
-
-	private void updateShield() {
-		if (isShielded) {
-			framesWithShieldActivated++;
-		}
-		else if (shieldShouldBeActivated()) {
-			activateShield();
-			shieldImage.showShield();
-		}
-		if (shieldExhausted()) {
-			deactivateShield();
-			shieldImage.hideShield();
-		}
-	}
-
-	private int getNextMove() {
-		int currentMove = movePattern.get(indexOfCurrentMove);
-		consecutiveMovesInSameDirection++;
-		if (consecutiveMovesInSameDirection == MAX_FRAMES_WITH_SAME_MOVE) {
-			Collections.shuffle(movePattern);
-			consecutiveMovesInSameDirection = 0;
-			indexOfCurrentMove++;
-		}
-		if (indexOfCurrentMove == movePattern.size()) {
-			indexOfCurrentMove = 0;
-		}
-		return currentMove;
-	}
-
-	private boolean bossFiresInCurrentFrame() {
-		return Math.random() < BOSS_FIRE_RATE;
-	}
-
-	private double getProjectileInitialPosition() {
-		return getLayoutY() + getTranslateY() + PROJECTILE_Y_POSITION_OFFSET;
-	}
-
-	private boolean shieldShouldBeActivated() {
-		return Math.random() < BOSS_SHIELD_PROBABILITY;
-	}
-
-	private boolean shieldExhausted() {
-		//System.out.println("SHIELD EXHAUSTED");
-		return framesWithShieldActivated == MAX_FRAMES_WITH_SHIELD;
-	}
-
-	private void activateShield() {
-		isShielded = true;
-		System.out.println("ACTIVATE SHIELD");
-	}
-
-	private void deactivateShield() {
-		isShielded = false;
-		System.out.println("DEACTIVATE SHIELD");
-		framesWithShieldActivated = 0;
-	}
-
-	public ShieldImage getShieldImage() {
-		return shieldImage;
-	}
-
-}
