+++ CW2024-master/CW2024-master/src/main/java/dev/ccr/dmscw2024/bosses/ISDProjectile.java
@@ -1,4 +1,6 @@
-package dev.ccr.dmscw2024;
+package dev.ccr.dmscw2024.bosses;
+
+import dev.ccr.dmscw2024.fundamentals.Projectile;
 
 public class ISDProjectile extends Projectile {
 
@@ -7,13 +9,16 @@ public class ISDProjectile extends Projectile {
     private static final int HORIZONTAL_VELOCITY = -15;
     private static final int INITIAL_X_POSITION = 950;
 
-    public ISDProjectile(double initialYPos) {
+    public ISDProjectile(double initialYPos, double projectileYPos) {
         super(IMAGE_NAME, IMAGE_HEIGHT, INITIAL_X_POSITION, initialYPos);
     }
 
     @Override
     public void updatePosition() {
         moveHorizontally(HORIZONTAL_VELOCITY);
+        if (getLayoutX() < 100) {
+            destroy();
+        }
     }
 
     @Override
