+++ CW2024-master/CW2024-master/src/main/java/dev/ccr/dmscw2024/specials/shield/Shield.java
@@ -4,4 +4,7 @@ public interface Shield {
     void showShield();
     void hideShield();
     void setPosition(double x, double y);
+
+    double getFitWidth();
+    double getFitHeight();
 }
